Authentication Commit

+++ controllers/goalController.js
@@ -1,11 +1,12 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 //@desc  Get goals
 //@route  GET /api/goals
 //@access  Private
 const getGoals = asyncHandler(async(req,res)=>{
-  const goals = await Goal.find()
+  const goals = await Goal.find({user: req.user.id})
      res.status(200).json(goals);
 })
 
@@ -18,7 +19,8 @@ const setGoals =  asyncHandler(async(req, res) => {
     throw new Error('Please Add a text field')
   }
   const goal = await Goal.create({
-    text: req.body.text
+    text: req.body.text,
+    user: req.user.id
   })
   res.status(200).json(goal);
 })
@@ -33,6 +35,19 @@ const updateGoals = asyncHandler(async(req, res) => {
   if(!goal) {
     res.status(400)
     throw new Error('Goal not Found')
+  }
+
+  const user = await User.findById(req.user.id)
+
+  // Check for user
+  if(!user) {
+    res.status(401)
+    throw new Error('User Not Found')
+  }
+
+  if(goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error('User Not Authorized')
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -50,6 +65,19 @@ const deleteGoals = asyncHandler(async(req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not Found");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  // Check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User Not Found");
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User Not Authorized");
   }
 
   const deletedGoal = await Goal.deleteOne({_id:req.params.id})
